+++ appmarket.py
@@ -48,9 +48,9 @@ def get_local_app_version(app_path):
 
 def main():
     if len(sys.argv) >= 2:
-        if sys.argv[1] >= "2.3.0":
+        if sys.argv[1] >= "2.3.5":
             # Initialize with splash screen
-            ekernel.splashScreen("ProcyonCLS AppMarket", "Version 2.3.0")
+            ekernel.splashScreen("ProcyonCLS AppMarket", "Version 2.3.5")
             
             while True:
                 # Display menu