+++ evaluator.py
@@ -9,9 +9,9 @@ term = Terminal()
 
 def main():
     if len(sys.argv) >= 2:
-        if sys.argv[1] >= "2.3.0":
+        if sys.argv[1] >= "2.3.5":
             # Initialize with splash screen
-            ekernel.splashScreen("ProcyonCLS Evaluator", "Version 2.3.0")
+            ekernel.splashScreen("ProcyonCLS Evaluator", "Version 2.3.5")
             
             while True:
                 # Display interface